rag/requirements.txt
clickhouse-connect
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
psycopg2-binary
python-crontab
python-dotenv
sentence_transformers